Fix implicit integer conversion warnings with clang-21 (#33)
Clang-21 is stricter about implicit integer conversions than clang-20.
The bitwise NOT operator (~) on uint8_t/uint16_t promotes to int, requiring
explicit casts back to the smaller types.

Fixed in 4 locations:
- Flags::operator~(): Added static_cast<std::uint8_t>
- Alu::sub8(): Added static_cast<std::uint8_t> for ~rhs
- Alu::sub16(): Added static_cast<std::uint16_t> for ~rhs
- Alu::sbc16(): Added static_cast<std::uint16_t> for ~rhs

All tests pass with clang-21.1.0.

+++ z80/common/include/z80/common/Flags.hpp
@@ -17,7 +17,7 @@ public:
   constexpr Flags operator&(const Flags rhs) const { return Flags(value_ & rhs.value_); }
   constexpr Flags operator|(const Flags rhs) const { return Flags(value_ | rhs.value_); }
   constexpr Flags operator^(const Flags rhs) const { return Flags(value_ ^ rhs.value_); }
-  constexpr Flags operator~() const { return Flags(~value_); }
+  constexpr Flags operator~() const { return Flags(static_cast<std::uint8_t>(~value_)); }
 
   [[nodiscard]] static constexpr Flags Carry() { return Flags(Flag::carry); }
   [[nodiscard]] static constexpr Flags Subtract() { return Flags(Flag::subtract); }
